pulling fom master

+++ tools/utils/seed/clean.ts
@@ -1,5 +1,8 @@
 import * as util from 'gulp-util';
+import { relative, normalize } from 'path';
 import * as rimraf from 'rimraf';
+
+import Config from '../../config';
 
 /**
  * Cleans the given path(s) using `rimraf`.
@@ -16,17 +19,24 @@ export function clean(paths: string|string[]): (done: () => void) => void {
 
     let promises = pathsToClean.map(p => {
       return new Promise(resolve => {
-        rimraf(p, e => {
-          if (e) {
-            util.log('Clean task failed with', e);
-          } else {
-            util.log('Deleted', util.colors.yellow(p || '-'));
-          }
-          resolve();
-        });
+        const relativePath: string = relative(Config.PROJECT_ROOT, p);
+        if (relativePath.startsWith('..')) {
+          util.log(util.colors.bgRed.white(`Cannot remove files outside the project root but tried "${normalize(p)}"`));
+          process.exit(1);
+        } else {
+          rimraf(p, e => {
+            if (e) {
+              util.log('Clean task failed with', e);
+            } else {
+              util.log('Deleted', util.colors.yellow(p || '-'));
+            }
+            resolve();
+          });
+        }
       });
     });
-    Promise.all(promises).then(() => done());
+    Promise.all(promises).then(() => (done || (() => 1))())
+      .catch(e => util.log(util.colors.red(`Error while removing files "${[].concat(paths).join(', ')}", ${e}`)));
   };
 }
